Add terms of service checkbox to signup. Change languagelink to a service.
