+++ timApp/defaultconfig.py
@@ -33,7 +33,7 @@ ACCESSIBILITY_STATEMENT_DOC = "tim/saavutettavuusseloste"
 
 # Path to TIM document containing the Terms of service The link to the document is shown in page footer.
 # If None, link to the document is not shown
-TERMS_OF_SERVICE_DOC = None
+TERMS_OF_SERVICE_DOC = "tim/kayttoehdot"
 
 ALLOWED_DOCUMENT_UPLOAD_MIMETYPES = ["text/plain"]
 COMPRESS_DEBUG = True
