+++ timApp/i18n/messages.sv.xlf
@@ -10824,6 +10824,22 @@ Separate multiple addresses with commas or write each address on a new line.</ta
           <context context-type="linenumber">120,121</context>
         </context-group>
       </trans-unit>
+      <trans-unit id="1876317285885218816" datatype="html">
+        <source>If you do not receive the message within 5 minutes, verify that your email address is entered correctly above. Additionally, check your Spam or Junk folder in case the message was filtered.</source>
+        <target state="new">If you do not receive the message within 5 minutes, verify that your email address is entered correctly above. Additionally, check your Spam or Junk folder in case the message was filtered.</target>
+        <context-group purpose="location">
+          <context context-type="sourcefile">static/scripts/tim/user/login-dialog.component.ts</context>
+          <context context-type="linenumber">209,210</context>
+        </context-group>
+      </trans-unit>
+      <trans-unit id="6518048135226350997" datatype="html">
+        <source><x id="TAG_INPUT" ctype="x-input" equiv-text="&quot;checkbox&quot; name=&quot;agree-checkbox&quot; [(ngModel)]=&quot;agreeToTerms&quot; [disabled]=&quot;emailSent&quot;&gt;"/> I have read and agree to the <x id="START_LINK" ctype="x-a" equiv-text="&lt;a href=&quot;{{linkLang.getLangLink(&apos;/view/&apos; + termsOfServicePath)}}&quot;&gt;"/>Terms of Service.<x id="CLOSE_LINK" ctype="x-a" equiv-text="&lt;/a&gt;"/></source>
+        <target state="new"><x id="TAG_INPUT" ctype="x-input" equiv-text="&quot;checkbox&quot; name=&quot;agree-checkbox&quot; [(ngModel)]=&quot;agreeToTerms&quot; [disabled]=&quot;emailSent&quot;&gt;"/> I have read and agree to the <x id="START_LINK" ctype="x-a" equiv-text="&lt;a href=&quot;{{linkLang.getLangLink(&apos;/view/&apos; + termsOfServicePath)}}&quot;&gt;"/>Terms of Service.<x id="CLOSE_LINK" ctype="x-a" equiv-text="&lt;/a&gt;"/></target>
+        <context-group purpose="location">
+          <context context-type="sourcefile">static/scripts/tim/user/login-dialog.component.ts</context>
+          <context context-type="linenumber">225,227</context>
+        </context-group>
+      </trans-unit>
     </body>
   </file>
 </xliff>
